rating: clear the held-back mechanical findings (battery round 1)
Grouped mechanical cleanups, no behavioral money change beyond the event_count width:

- event_count widened to BIGINT end to end (the earlier fast-follow's widening didn't
  carry into the YAML-rater rewrite): the SQL COUNT(*) cast (::int -> ::bigint), the
  rated_usage column (0002 SQL + rating Alembic + the integration schema), and the e2e
  scan (int -> int64). An INTEGER column silently caps a hot (auth, model, hour) bucket
  at 2^31 while SUM(event_count) is already ::bigint.
- Deleted the dead ErrDerivationChain sentinel: it was documented as returned by the
  oracle but never was. One-hop derivation is enforced at LOAD (buildPriceBook rejects
  multi-hop / dangling derived_from) and in ResolveEvent, so a deeper chain can never
  reach the oracle — there is nothing to return. Replaced with a comment stating where
  the invariant actually lives.
- Renamed the overclaiming TestConformance_SQLModelMatchesRateOracle ->
  TestOracleModel_SelfConsistent: it runs NO SQL, it pins the in-Go oracle's
  self-consistency; the REAL SQL conformance is the integration test (cross-reference
  fixed).
- TestLoad_MalformedYAMLFailsClosed: the doc now accurately enumerates every shape it
  pins, INCLUDING the inconsistent-premium-policy cases (multiplier-no-factor,
  markup-with-factor, unknown-policy) it already tested but didn't mention.
- Corrected the stale "pointer-not-copy rule" comment on TestLoad_FineTuneIdentityPremium
  (it tests identity-default = base exactly, not the propagation rule).
- Fixed RateResult.TotalCost doc: the SQL COALESCEs SUM to 0, so an empty window returns
  "0", never "" — doc now matches reality.
- migrations/README: the fine-tune base-linkage "gap (flagged)" is now "closed" (carried
  on billing_event.base_model).

+++ internal/e2e/e2e_test.go
@@ -471,7 +471,7 @@ func TestE2E_StreamedRequestBecomesMoney(t *testing.T) {
 		nRollups                                     int
 		ruAuthID, ruModelID, cost                    string
 		ruPrompt, ruCached, ruCompletion, ruBillable int64
-		eventCount                                   int
+		eventCount                                   int64 // BIGINT column
 	)
 	if err := h.db.QueryRow("SELECT COUNT(*) FROM rated_usage").Scan(&nRollups); err != nil {
 		t.Fatalf("count rated_usage: %v", err)
